Imgur OAuth Login Complete
If we don't have an access token, then show a Login button.  Clicking the login button loads the Imgur OAuth screen, which gets an access token (this access token is stored in Shared Preferences for later).

With an access token, we make a request to the Imgur API for the user's images, but right now we show DummyItems instead of the images.