Fixed AppBar Visibility Issues
The AppBar would sometimes be hidden on the single image view, especially after scrolling down in the image gallery.  Because there was no scrolling in the single image view, the AppBar position would be stuck in whatever state it was from the image gallery.  I also updated the layout parameters on the image gallery view so that the AppBar snaps either visible or hidden, instead of sometimes showing in a half visible state.